model

+++ onef_core/model/model.go
@@ -1,5 +1,5 @@
 package model
 
 type Model struct {
-	Id int `json:"id,omitempty" gorm:"column:id;primary_key;auto_increment" bson:"id"`
+	Id int64 `json:"id,omitempty" gorm:"column:id;primary_key;auto_increment" bson:"id"`
 }
